Add an option to vs_core to allow frame skipping
If this option is enabled, the image you obtain by calling dequeue_image in a
plugin is always the latest image obtained from the camera.

This allows to simplify the creation of plugin that do not work in real time
and need to create one thread to compute and keep discarding image in the
loop_fct. This technique is useless if the option is enabled

To enable the option simply add:
[SkipFrames] 1
to vs_core.conf